refactor(core): reorganize directory structure and update dependencies
- Moved process-related files to internal/core/process package
- Moved config file to internal/core/config package
- Moved application files to internal/core/app package
- Added github.com/stretchr/testify dependency
- Documented logging choice (Zap over Logrus) in CURRENT_STATUS.md
