+++ go.mod
@@ -8,4 +8,5 @@ require (
 	github.com/spf13/cobra v1.8.0
 	github.com/spf13/viper v1.18.2
 	go.uber.org/zap v1.27.0
+	github.com/stretchr/testify v1.8.4
 )
