+++ internal/core/app/application.go
@@ -1,4 +1,4 @@
-package core
+package app
 
 import (
 	"context"
