+++ internal/core/process/orchestrator_integration_test.go
@@ -1,4 +1,4 @@
-package core
+package process
 
 import (
 	"fmt"
