+++ internal/core/process/orchestrator_management.go
@@ -1,4 +1,4 @@
-package core
+package process
 
 import (
 	"fmt"
