+++ internal/core/process/orchestrator_output.go
@@ -1,4 +1,4 @@
-package core
+package process
 
 import (
 	"bytes"
