+++ internal/core/process/orchestrator_process.go
@@ -1,4 +1,4 @@
-package core
+package process
 
 import (
 	"fmt"
