+++ internal/core/process/process_types.go
@@ -1,4 +1,4 @@
-package core
+package process
 
 import (
 	"fmt"
